reorg mh5_controller
